refactor: move some defaults helpers out of utils
tl;dr: there's a circular import here between utils.ts and defaults.ts. in the
following commits this results in a build error, so i'm making utils.ts no
longer depend on defaults.ts.
